agents/ten_packages/extension/vision_tool/requirements.txt
colorlog
pillow